requirements.txt
django==3.0.7
django-debug-toolbar==2.2
dj-database-url==0.5.0
Pillow==7.1.2
environs==8.0.0
djangorestframework==3.12.1
requests==2.24.0
geopy==2.0.0
django-admin-sortable2==0.7.7
